visualizer: fixed includes after move files
