+++ src/inet/visualizer/osg/base/LinkOsgVisualizerBase.cc
@@ -22,7 +22,7 @@
 
 #include "inet/common/LayeredProtocolBase.h"
 #include "inet/common/ModuleAccess.h"
-#include "inet/common/OsgScene.h"
+#include "inet/visualizer/osg/util/OsgScene.h"
 #include "inet/mobility/contract/IMobility.h"
 
 namespace inet {
