+++ src/inet/visualizer/osg/common/PacketDropOsgVisualizer.cc
@@ -18,8 +18,8 @@
 #include "inet/visualizer/osg/common/PacketDropOsgVisualizer.h"
 
 #include "inet/common/ModuleAccess.h"
-#include "inet/common/OsgScene.h"
-#include "inet/common/OsgUtils.h"
+#include "inet/visualizer/osg/util/OsgScene.h"
+#include "inet/visualizer/osg/util/OsgUtils.h"
 
 namespace inet {
