+++ src/inet/visualizer/osg/common/StatisticOsgVisualizer.cc
@@ -18,7 +18,7 @@
 #include "inet/visualizer/osg/common/StatisticOsgVisualizer.h"
 
 #include "inet/common/ModuleAccess.h"
-#include "inet/common/OsgUtils.h"
+#include "inet/visualizer/osg/util/OsgUtils.h"
 
 namespace inet {
