+++ src/inet/visualizer/osg/linklayer/InterfaceTableOsgVisualizer.cc
@@ -18,7 +18,7 @@
 #include "inet/visualizer/osg/linklayer/InterfaceTableOsgVisualizer.h"
 
 #include "inet/common/ModuleAccess.h"
-#include "inet/common/OsgUtils.h"
+#include "inet/visualizer/osg/util/OsgUtils.h"
 
 namespace inet {
