+++ src/inet/visualizer/osg/mobility/MobilityOsgVisualizer.cc
@@ -24,8 +24,8 @@
 #include <osgText/Text>
 
 #include "inet/common/ModuleAccess.h"
-#include "inet/common/OsgScene.h"
-#include "inet/common/OsgUtils.h"
+#include "inet/visualizer/osg/util/OsgScene.h"
+#include "inet/visualizer/osg/util/OsgUtils.h"
 
 namespace inet {
