+++ src/inet/visualizer/osg/networklayer/RoutingTableOsgVisualizer.cc
@@ -20,8 +20,8 @@
 #include <osg/Geode>
 #include <osg/LineWidth>
 
-#include "inet/common/OsgScene.h"
-#include "inet/common/OsgUtils.h"
+#include "inet/visualizer/osg/util/OsgScene.h"
+#include "inet/visualizer/osg/util/OsgUtils.h"
 
 namespace inet {
