+++ src/inet/visualizer/osg/physicallayer/MediumOsgVisualizer.cc
@@ -28,8 +28,8 @@
 #include <osgDB/ReadFile>
 
 #include "inet/common/ModuleAccess.h"
-#include "inet/common/OsgScene.h"
-#include "inet/common/OsgUtils.h"
+#include "inet/visualizer/osg/util/OsgScene.h"
+#include "inet/visualizer/osg/util/OsgUtils.h"
 
 #ifdef INET_WITH_PHYSICALLAYERWIRELESSCOMMON
 #include "inet/physicallayer/wireless/common/pathloss/FreeSpacePathLoss.h"
