+++ src/inet/visualizer/osg/power/EnergyStorageOsgVisualizer.h
@@ -19,7 +19,7 @@
 #define __INET_ENERGYSTORAGEOSGVISUALIZER_H
 
 #include "inet/common/ModuleRefByPar.h"
-#include "inet/common/OsgUtils.h"
+#include "inet/visualizer/osg/util/OsgUtils.h"
 #include "inet/visualizer/base/EnergyStorageVisualizerBase.h"
 #include "inet/visualizer/osg/scene/NetworkNodeOsgVisualizer.h"
