+++ src/inet/visualizer/osg/scene/NetworkNodeOsgVisualizer.cc
@@ -20,8 +20,8 @@
 #include <omnetpp/osgutil.h>
 
 #include "inet/common/ModuleAccess.h"
-#include "inet/common/OsgScene.h"
-#include "inet/common/OsgUtils.h"
+#include "inet/visualizer/osg/util/OsgScene.h"
+#include "inet/visualizer/osg/util/OsgUtils.h"
 
 namespace inet {
