+++ src/inet/visualizer/osg/scene/SceneOsgEarthVisualizer.cc
@@ -18,8 +18,8 @@
 #include "inet/visualizer/osg/scene/SceneOsgEarthVisualizer.h"
 
 #include "inet/common/ModuleAccess.h"
-#include "inet/common/OsgScene.h"
-#include "inet/common/OsgUtils.h"
+#include "inet/visualizer/osg/util/OsgScene.h"
+#include "inet/visualizer/osg/util/OsgUtils.h"
 #include "inet/environment/contract/IPhysicalEnvironment.h"
 #include "inet/mobility/contract/IMobility.h"
 #include "inet/visualizer/osg/scene/NetworkNodeOsgVisualizer.h"
