+++ src/inet/visualizer/osg/transportlayer/TransportConnectionOsgVisualizer.cc
@@ -18,7 +18,7 @@
 #include "inet/visualizer/osg/transportlayer/TransportConnectionOsgVisualizer.h"
 
 #include "inet/common/ModuleAccess.h"
-#include "inet/common/OsgUtils.h"
+#include "inet/visualizer/osg/util/OsgUtils.h"
 
 namespace inet {
