+++ src/inet/visualizer/osg/util/OsgScene.cc
@@ -15,7 +15,7 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 //
 
-#include "inet/common/OsgScene.h"
+#include "inet/visualizer/osg/util/OsgScene.h"
 
 namespace inet {
